+++ test/matchBrackets.test.ts
@@ -1,18 +1,23 @@
 import * as assert from 'assert';
+import {Decoration} from '@codemirror/view';
 import {syntaxTree} from '@codemirror/language';
 import {javascript} from '@codemirror/lang-javascript';
 import {css} from '@codemirror/lang-css';
 import {json} from '@codemirror/lang-json';
 import lua from '../src/lua';
+import {exclude} from '../src/javascript';
 import {
 	findEnclosingBrackets,
 	findEnclosingPlainBrackets,
 	trySelectMatchingBrackets,
 	selectMatchingBrackets,
+	bracketDeco,
 } from '../src/matchBrackets';
-import {createState} from './util';
-import type {LanguageSupport, Config, MatchResult} from '@codemirror/language';
-import type {Selection} from '../src/matchBrackets';
+import {createState, convertRangeSet} from './util';
+import type {Range} from '@codemirror/state';
+import type {LanguageSupport, MatchResult} from '@codemirror/language';
+import type {Selection, RequiredConfig} from '../src/matchBrackets';
+import type {DocRange} from '../src/fold';
 
 declare type Result = [number, number];
 
@@ -20,9 +25,16 @@ const javascriptLanguage = javascript(),
 	cssLanguage = css(),
 	jsonLanguage = json(),
 	luaLanguage = lua(),
-	config: Config = {
+	mark = Decoration.mark({}),
+	toRange = ({from, to}: DocRange): Range<Decoration> => mark.range(from, to),
+	config: RequiredConfig = {
 		brackets: '()[]{}',
 		maxScanDistance: 1e4,
+		afterCursor: true,
+		exclude,
+		renderMatch({start, end}) {
+			return [toRange(start), ...end ? [toRange(end)] : []];
+		},
 	};
 
 const mockTest = (bracket: MatchResult | null | undefined, result?: Result | null): void => {
@@ -33,10 +45,10 @@ const mockTest = (bracket: MatchResult | null | undefined, result?: Result | nul
 	},
 	lezerTest = (doc: string, lang: LanguageSupport, pos: number, result?: Result): void => {
 		const node = syntaxTree(createState(doc, lang)).resolveInner(pos, -1);
-		mockTest(findEnclosingBrackets(node, pos, config.brackets!), result);
+		mockTest(findEnclosingBrackets(node, pos, config.brackets), result);
 	},
 	plainTest = (doc: string, lang: LanguageSupport | undefined, pos: number, result: Result | null): void => {
-		mockTest(findEnclosingPlainBrackets(createState(doc, lang), pos, config as Required<Config>), result);
+		mockTest(findEnclosingPlainBrackets(createState(doc, lang), pos, config), result);
 	},
 	trySelectTest = (doc: string, pos: number, assoc: 1 | -1, inside: boolean, result: Selection | false): void => {
 		assert.deepStrictEqual(
@@ -49,6 +61,14 @@ const mockTest = (bracket: MatchResult | null | undefined, result?: Result | nul
 		assert.deepStrictEqual(
 			selectMatchingBrackets(createState(doc, []), pos),
 			result,
+			`pos: ${pos}`,
+		);
+	},
+	decoTest = (doc: string, pos: number, result: number[] = []): void => {
+		const state = createState(doc, javascriptLanguage, [pos]);
+		assert.deepStrictEqual(
+			convertRangeSet(bracketDeco(state, config), doc.length),
+			result.map(i => [i, i + 1]),
 			`pos: ${pos}`,
 		);
 	};
@@ -126,5 +146,49 @@ describe('select bracket pair on both sides', () => {
 	it('inside', () => {
 		selectTest(' [text] ', 2, {anchor: 2, head: 6});
 		selectTest(' [text] ', 6, {anchor: 6, head: 2});
+	});
+});
+
+describe('bracket decorations', () => {
+	it('before bracket', () => {
+		decoTest('[true]', 0, [0, 5]);
+		decoTest('[true]', 5, [0, 5]);
+		decoTest('[true', 0, [0]);
+		decoTest('"[x]"', 1, [1, 3]);
+		decoTest('"[x]"', 3, [1, 3]);
+		decoTest('"[x"', 1, [1]);
+		decoTest('"x]"', 2, [2]);
+		decoTest('("[x")', 2, [2]);
+	});
+	it('after bracket', () => {
+		decoTest('[true]', 1, [0, 5]);
+		decoTest('[true]', 6, [0, 5]);
+		decoTest('[true', 1, [0]);
+		decoTest('"[x]"', 2, [1, 3]);
+		decoTest('"[x]"', 4, [1, 3]);
+		decoTest('"[x"', 2, [1]);
+		decoTest('"x]"', 3, [2]);
+		decoTest('("[x")', 3, [2]);
+	});
+	it('inside brackets', () => {
+		decoTest('[true]', 2, [0, 5]);
+		decoTest('"[foo]"', 3, [1, 5]);
+		decoTest('("[foo")', 4, [0, 7]);
+	});
+	it('no brackets', () => {
+		decoTest('[true', 2);
+		decoTest('"[foo"', 3);
+	});
+	it('RegExp literal', () => {
+		decoTest('/[x]/', 1);
+		decoTest('/[x]/', 2);
+		decoTest('/[x]/', 3);
+		decoTest('/[x]/', 4);
+		decoTest('/[foo]/', 3);
+		decoTest('(/[x]/)', 2, [0, 6]);
+		decoTest('(/[x]/)', 3, [0, 6]);
+		decoTest('(/[x]/)', 4, [0, 6]);
+		decoTest('(/[x]/)', 5, [0, 6]);
+		decoTest('(/[foo]/)', 4, [0, 8]);
 	});
 });